Consolidate our handling of JSON types.
We had two separate implementations of JSON type parsing, and that just
seems wasteful. One was similar to cty's, and used a custom JSON decoder
to parse types. One was a novel algorithm that decoded the JSON bytes
into an `interface{}` and read the resulting Go types using a switch
statement to understand the type information.

This consolidates to only use the cty-esque implementation, which is
presumably more accurate and stable.
